% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convert degrees Celsius to degrees Fahrenheit}
\usage{
celsius_to_fahrenheit(dc)
}
\arguments{
\item{dc}{Temperature in degrees Celsius}
}
\value{
Temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius to degrees Fahrenheit
}
\examples{
celsius_to_fahrenheit(0)
celsius_to_fahrenheit(100)
}
