% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert degrees Fahrenheit to degrees Celsius}
\usage{
fahrenheit_to_celsius(df)
}
\arguments{
\item{df}{Temperature in degrees Fahrenheit}
}
\value{
Temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit to degrees Celsius
}
\examples{
fahrenheit_to_celsius(32)
fahrenheit_to_celsius(212)
}
