% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kilogram_to_pound.R
\name{kilogram_to_pound}
\alias{kilogram_to_pound}
\title{Convert kilograms to pounds}
\usage{
kilogram_to_pound(kg)
}
\arguments{
\item{kg}{Weight in kilograms}
}
\value{
Weight in pounds
}
\description{
Convert kilograms to pounds
}
\examples{
kilogram_to_pound(100)
kilogram_to_pound(300)
}
