% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pound_to_kilogram.R
\name{pound_to_kilogram}
\alias{pound_to_kilogram}
\title{Convert pounds to kilograms}
\usage{
pound_to_kilogram(p)
}
\arguments{
\item{p}{Weight in pounds}
}
\value{
Weight in kilograms
}
\description{
Convert pounds to kilograms
}
\examples{
pound_to_kilogram(100)
pound_to_kilogram(300)
}
