% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(dc)
}
\arguments{
\item{dc}{double}
}
\value{
double
}
\description{
Temperature Conversion: Celsius to Fahrenheit
}
\examples{
celsius_to_fahrenheit(0)

}
