% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(df)
}
\arguments{
\item{df}{double}
}
\value{
double
}
\description{
Temperature Conversion: Fahrenheit to Celsius
}
\examples{
fahrenheit_to_celsius(32)

}
