% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kilogram_to_pound.R
\name{kilogram_to_pound}
\alias{kilogram_to_pound}
\title{Kilogram to Pound}
\usage{
kilogram_to_pound(kg)
}
\arguments{
\item{kg}{double}
}
\value{
double
}
\description{
Weight Conversion: Kilograms to Pounds
}
\examples{
kilogram_to_pound(60)

}
