% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pound_to_kilogram.R
\name{pound_to_kilogram}
\alias{pound_to_kilogram}
\title{Pound to Kilogram}
\usage{
pound_to_kilogram(p)
}
\arguments{
\item{p}{double}
}
\value{
double
}
\description{
Weight Conversion: Pound to Kilogram
}
\examples{
pound_to_kilogram(20)

}
